requirements.txt
Django>=1.11
gunicorn>=19.7.0
psycopg2==2.7.1
dj-database-url==0.4.2
celery>=4.0.2
gevent>=1.2.1
whitenoise>=3.3.0
djangorestframework>=3.6.2
bleach>=2.0.0
Markdown>=2.6.8
python-dotenv==0.6.1
boto3
django-queued-storage>=0.8.0
django-storages>=1.5.2
mixpanel>=4.3.2
